作成したユーザーのみ削除機能実行可能_#8

+++ app/controllers/api/v1/articles_controller.rb
@@ -1,7 +1,7 @@
 module Api
   module V1
     class ArticlesController < ApplicationController
-      skip_before_action :authenticate_user
+      skip_before_action :authenticate_user, only: [:index,:show,:destroy]
 
       def index
         articles = Article.all.as_json(include:[:user, :favorites])
@@ -24,9 +24,7 @@ module Api
             articles: article,
           }, status: :ok
         else
-          render json: {
-            articles: article.error,
-          }
+          render json: {status: :bad_request}
         end
         # pp FirebaseIdToken::Signature.verify(params.require(:headers).permit(:Authorization)[:Authorization])
       end
@@ -38,41 +36,42 @@ module Api
             articles: article,
           },status: :ok
         else
-          render json: {
-            articles: article.error
-          }
+          render json: {status: :bad_request}
         end
       end
 
       def destroy
+        raise ArgumentError, 'BadRequest Parameter' if payload.blank?
         article = Article.find(params[:id])
-        article.destroy
+        if article.user_id == current_user.id
+          article.destroy
+          render json: {status: :ok}
+        else
+          render json: {status: :bad_request}
+        end
       end
 
       private
 
-    def set_id_params
-      params.require(:article).permit(:id)
-    end
+      def set_id_params
+        params.require(:article).permit(:id)
+      end
 
-    def article_params
-      params.require(:article).permit(:title,:text)
-    end
+      def article_params
+        params.require(:article).permit(:title,:text)
+      end
 
-    def token
-      params.require(:headers).permit(:Authorization)[:Authorization]
-    end
+      def token_from_request_headers
+        params.require(:headers).permit(:Authorization)[:Authorization]
+      end
 
-    def payload
-      # https://github.com/penguinwokrs/firebase-auth-rails
-      # tokenの検証が成功すれば、@payloadに入る
-      # https://github.com/fschuindt/firebase_id_token
-      # documents
-      # https://www.rubydoc.info/gems/firebase_id_token
-      @payload ||= FirebaseIdToken::Signature.verify token
-    end
+      def token
+        params[:token] || token_from_request_headers
+      end
 
-
+      def payload
+        @payload ||= FirebaseIdToken::Signature.verify token
+      end
     end
   end
 end
